Fix processing of HTTP redirects
For some reason, the nginx proxy decided to cache the HTTP 302 redirect
response sent back by the image registry. The reason could have been
that the Expires header was not set enough into the past but I didn't
investigate in detail.

The caching of the HTTP redirect caused the proxy to break. The proxy
cached the HTTP 302 reponse and then in @handle_redirects it found this
response in cache (cache HIT in the logs) and returned it to the client.

The error message seen in the logs caused by this issue:
[error] 78#78: *46 invalid URL prefix in ""