add option to change cache behavior (CACHE_GREEDY_MODE)
In default mode a layer exist only once in the cache, and regardless
which repository it is initially from, each repository get a hit on
this layer now.
Pitfall: You can not push images that reuse layers from a different
registry (`ALLOW_PUSH=true`).
Setting this option to 'false' adds the host to the cache key to solve this.

Default is `true` to not change default behavior.
WARNING: Changing this setting invalidates your cache!